@@ -1,3 +1,71 @@
+[tool.poetry]
+authors = ["Your Name <you@example.com>"]
+description = ""
+name = "openstreetmap-ng"
+packages = [
+  {include = "src"},
+]
+version = "0.0.0"
+
+[tool.poetry.dependencies]
+aioboto3 = "^12.1.0"
+aiosmtplib = "^3.0.1"
+alembic = "^1.13.1"
+anyio = "^4.2.0"
+argon2-cffi = "^23.1.0"
+arrow = "^1.3.0"
+asyncache = "^0.3.1"
+asyncpg = "^0.29.0"
+authlib = "^1.3.0"
+babel = "^2.14.0"
+beautifulsoup4 = "^4.12.2"
+bleach = "^6.1.0"
+cachetools = "^5.3.2"
+cryptography = "^41.0.7"
+cython = "^3.0.7"
+email-validator = "^2.1.0.post1"
+fastapi = "^0.108.0"
+feedgen = "^1.0.0"
+geoalchemy2 = "^0.14.3"
+httpx = {extras = ["brotli", "http2"], version = "^0.26.0"}
+humanize = "^4.9.0"
+jinja2 = "^3.1.2"
+lxml = "^5.0.1"
+markdown-it-py = {extras = ["plugins"], version = "^3.0.0"}
+msgspec = "^0.18.5"
+networkx = "^3.2.1"
+numpy = "^1.26.3"
+orjson = "^3.9.10"
+phonenumbers = "^8.13.27"
+pillow = "^10.2.0"
+psycopg2-binary = "^2.9.9"
+pycryptodome = "^3.19.1"
+pydantic = "^2.5.3"
+pyinstrument = "^4.6.1"
+pyproj = "^3.6.1"
+python = "^3.12"
+python-dateutil = "^2.8.2"
+python-magic = "^0.4.27"
+python-multipart = "^0.0.6"
+pytz = "^2023.3.post1"
+pyyaml = "^6.0.1"
+requests = "^2.31.0"
+rfc3986 = "^2.0.0"
+scikit-learn = "^1.3.2"
+setuptools = "^69.0.3"
+shapely = "^2.0.2"
+sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
+tqdm = "^4.66.1"
+uvicorn = {extras = ["standard"], version = "^0.25.0"}
+
+[tool.poetry.group.test.dependencies]
+pytest = "^7.4.4"
+xmltodict = "^0.13.0"
+
+[build-system]
+build-backend = "poetry.core.masonry.api"
+requires = ["poetry-core", "setuptools", "wheel", "Cython"]
+
 [tool.ruff]
 # Exclude a variety of commonly ignored directories.
 exclude = [
@@ -22,16 +90,18 @@ exclude = [
   "dist",
   "node_modules",
   "venv",
-  "alembic/versions",
-  "models/db/__init__.py",
+  "src/alembic/versions",
 ]
 
 indent-width = 4
 line-length = 120
-target-version = "py311"
+target-version = "py312"
 
 [tool.ruff.lint]
-ignore = []
+ignore = [
+  "S101", # assert
+  "PT011", # pytest-raises-too-broad
+]
 select = [
   # see https://docs.astral.sh/ruff/rules/ for rules documentation
   "A", # flake8-builtins
@@ -94,7 +164,7 @@ skip-magic-trailing-comma = false
 line-ending = "lf"
 
 [tool.ruff.lint.flake8-builtins]
-builtins-ignorelist = ["id", "open", "type"]
+builtins-ignorelist = ["id", "input", "open", "type"]
 
 [tool.ruff.lint.flake8-gettext]
 extend-function-names = ["pgettext", "npgettext"]
